requirements-openai.txt
# OpenAI Comprehensive Integration Requirements
# Author: Nik Jois
# Email: nikjois@llamasearch.ai

# Core OpenAI SDK
openai>=1.51.0

# Audio processing for transcription and TTS
pydub>=0.25.1
soundfile>=0.12.1
librosa>=0.10.1
scipy>=1.11.0

# Image processing for DALL-E integration
Pillow>=10.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0

# Data processing and analysis
numpy>=1.24.0
pandas>=2.0.0

# File handling and storage
aiofiles>=23.2.0
boto3>=1.34.0  # For S3-compatible storage
minio>=7.2.0

# Redis for caching and session management
redis>=5.0.0
hiredis>=2.2.0

# PostgreSQL for data persistence
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# Background task processing
celery>=5.3.0
kombu>=5.3.0

# Monitoring and observability
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-requests>=0.42b0
opentelemetry-exporter-prometheus>=1.12.0
jaeger-client>=4.8.0

# Security and authentication
cryptography>=41.0.0
jwt>=1.3.1
bcrypt>=4.1.0

# Configuration management
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# HTTP and networking
httpx>=0.25.0
aiohttp>=3.9.0
websockets>=12.0

# Utilities
python-multipart>=0.0.6  # For file uploads
strenum>=0.4.15  # Better enum support
python-slugify>=8.0.0
click>=8.1.0  # Command line interface

# Development and testing (optional in production)
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
faker>=20.0.0 